Spark support: Rename with-spark-schnorr  example to with-spark
