--- examples/with-spark-schnorr/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "noEmit": true,
-    "tsBuildInfoFile": "./.cache/.tsbuildinfo"
-  },
-  "include": ["src/**/*.ts", "src/**/*.js", "src/**/*.json"]
-}
