+++ examples/with-spark/.env.local.example
@@ -1,0 +1,30 @@
+# Spark wallet configuration
+# BIP39 mnemonic to restore an existing wallet.
+# Leave blank to auto-generate one on first run.
+MNEMONIC=
+
+# Network: MAINNET or REGTEST (default: REGTEST)
+SPARK_NETWORK=REGTEST
+
+# Optional: Spark address to send a test transfer to (for regtest acceptance test)
+RECEIVER_SPARK_ADDRESS=
+
+# -------------------------------------------------------------------
+# Optional: Turnkey credentials for signature comparison
+# Set these if you want to compare native Spark SDK Schnorr signatures
+# against Turnkey's signRawPayload output for the same key.
+# -------------------------------------------------------------------
+BASE_URL=
+ORGANIZATION_ID=
+API_PUBLIC_KEY=
+API_PRIVATE_KEY=
+
+# The Turnkey address that holds the same private key
+# as the Spark identity key derived from the mnemonic above.
+TURNKEY_IDENTITY_ADDRESS=
+# Compressed 33-byte public key (hex) of the identity key held in Turnkey.
+# Used by TurnkeySparkSigner.getIdentityPublicKey() without a round-trip.
+IDENTITY_PUBLIC_KEY_HEX=
+
+# Token transfer amount to send (default: 1000)
+TRANSFER_AMOUNT=1000
