+++ examples/with-spark/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@turnkey/example-with-spark",
+  "private": true,
+  "version": "0.1.0",
+  "scripts": {
+    "start": "tsx src/index.ts",
+    "clean": "rimraf ./dist ./.cache",
+    "typecheck": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@buildonspark/issuer-sdk": "^0.1.23",
+    "@buildonspark/spark-sdk": "0.7.5",
+    "@noble/curves": "^1.9.7",
+    "@noble/hashes": "^1.4.0",
+    "@scure/bip39": "^1.3.0",
+    "@scure/btc-signer": "^1.5.0",
+    "@turnkey/sdk-server": "workspace:*",
+    "@turnkey/spark": "workspace:*",
+    "dotenv": "^16.0.3"
+  },
+  "devDependencies": {
+    "tsx": "^4.7.0",
+    "typescript": "5.4.3"
+  }
+}
